+++ feed/tests/root/conf.py
@@ -40,8 +40,8 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Sphinx Syndicate <Test>'
-copyright = u'2009, Dan MacKinlay'
+project = u'Sphinx Syndicate Test'
+copyright = u'2011, ReST wranglers'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -123,7 +123,7 @@ html_theme = 'default'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = []
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.
@@ -171,6 +171,7 @@ htmlhelp_basename = 'SphinxSyndicateTestdoc'
 
 siteurl = 'file://' + os.path.abspath('./_build/html') #won't work in windows
 feed_base_url = siteurl
-feed_description = "The outpourings of one Dan MacKinlay"
+feed_description = "The outpourings of a Sphinx developer"
 feed_filename = 'rss.xml'
+feed_title = 'Sphinx Syndicate Test Title'