+++ feed/tests/test_build.py
@@ -10,7 +10,7 @@ from StringIO import StringIO
 from util import test_root, raises, raises_msg, Struct,\
   ListOutput, TestApp, with_app, gen_with_app, path, with_tempdir,\
   write_file, sprint
-from sphinxcontrib.feed.path import path
+from path import path
 from BeautifulSoup import BeautifulSoup
 import feedparser
 from datetime import datetime
@@ -32,16 +32,6 @@ ENV_WARNINGS = "" #none at the moment. if needed, they will look like:
 
 FEED_WARNINGS = ENV_WARNINGS + "" # nothing here either yet
 
-TEMP_DOC = """:Date: 2002-08-11
-:Author: Santa
-
-blog post of irrelevant age
-===========================
-
-Has some cruft.
-
-"""
-
 def test_feed():
     app = TestApp(buildername='html', warning=feed_warnfile, cleanenv=True)  
     app.build(force_all=True, filenames=[]) #build_all misses the crucial finish signal
@@ -56,6 +46,7 @@ def test_feed():
     # see http://www.feedparser.org/
     f = feedparser.parse(rss_path)
     yield assert_equals, f.bozo, 0 #feedparser well-formedness detection. We want this.
+    yield assert_equals, f.feed['title'], 'Sphinx Syndicate Test Title'
     entries = f.entries
     yield assert_equals, entries[0].updated_parsed[0:6], (2001, 8, 11, 13, 0, 0)
     yield assert_equals, entries[0].title, "The latest blog post"
@@ -95,37 +86,3 @@ def test_feed():
     
     app.cleanup()
     app2.cleanup()
-
-# def test_feed_cleanup():
-#     """
-#     do it all again, but purge a file in between and see how we go
-#     """
-#     app0 = TestApp(buildername='html', warning=feed_warnfile, cleanenv=True)  
-#     srcdir = path(app0.srcdir)
-#     tmp_path = srcdir/'tmp.rst'
-#     rss_path = os.path.join(app0.outdir, 'rss.xml')
-#     app0.cleanup()
-#     del(app0)
-#     try:
-#         tmp_path_handle = tmp_path.open('w')
-#         tmp_path_handle.write(TEMP_DOC)
-#         tmp_path_handle.close()
-#         yield assert_true, tmp_path.exists()
-#         yield assert_equals, tmp_path.open().read(), TEMP_DOC
-#         # raise Exception
-#         app1 = TestApp(buildername='html', warning=feed_warnfile, cleanenv=True)  
-#         app1.build()
-#         f = feedparser.parse(rss_path)
-#         entries = f.entries
-#         yield assert_equals, entries[0].updated_parsed[0:6], (2002, 8, 11)
-#         yield assert_equals, entries[0].title, "blog post of irrelevant age"
-#         tmp_path.remove()
-#         yield assert_false, tmp_path.exists()
-#         app1 = TestApp(buildername='html', warning=feed_warnfile)  
-#         app1.build()
-#         f = feedparser.parse(rss_path)
-#         entries = f.entries
-#         yield assert_equals, entries[0].updated_parsed[0:6], (2001, 8, 11)
-#         yield assert_equals, entries[0].title, "The latest blog post"
-#     finally:
-#         pass #if tmp_path.exists(): tmp_path.remove()