+++ spelling/sphinxcontrib/spelling.py
@@ -6,25 +6,163 @@
 """Spelling checker extension for Sphinx.
 """
 
-from docutils.nodes import GenericNodeVisitor
+import codecs
+import collections
+import imp
+import itertools
+import os
+import re
+import textwrap
+import tempfile
+import xmlrpclib
+
 #from docutils import core
+from docutils.frontend import OptionParser
+from docutils.io import StringOutput
+import docutils.nodes
+from docutils.nodes import GenericNodeVisitor
+from docutils.parsers import rst
 from docutils.writers import Writer
 from sphinx.builders import Builder
 from sphinx.util.console import bold, darkgreen
-from docutils.frontend import OptionParser
+from sphinx.util.console import purple, red, darkgreen, darkgray
 from sphinx.util.nodes import inline_all_toctrees
-from docutils.io import StringOutput
-
-import itertools
-import re
-import textwrap
-import imp
 
 import enchant
-from enchant.tokenize import get_tokenizer
+from enchant.tokenize import (get_tokenizer, tokenize,
+                              Filter, EmailFilter, WikiWordFilter,
+                              unit_tokenize, wrap_tokenizer,
+                              )
 
-# TODO - Directive (or comment syntax?) to allow words to be ignored in a document
 # TODO - Words with multiple uppercase letters treated as classes and ignored
+
+class SpellingDirective(rst.Directive):
+    """Custom directive for passing instructions to the spelling checker.
+
+    .. spelling::
+
+       word1
+       word2
+    
+    """
+
+    option_spec = {}
+    has_content = True
+
+    def run(self):
+        env = self.state.document.settings.env
+        if not hasattr(env, 'spelling_document_filters'):
+            env.spelling_document_filters = collections.defaultdict(list)
+        good_words = []
+        for entry in self.content:
+            if not entry:
+                continue
+            good_words.extend(entry.split())
+        if good_words:
+            env.app.info('Extending local dictionary with %s' % str(good_words))
+            env.spelling_document_filters[env.docname].append(
+                IgnoreWordsFilterFactory(good_words)
+                )
+        return []
+
+
+class AcronymFilter(Filter):
+    """If a word looks like an acronym (all upper case letters),
+    ignore it.
+    """
+    def _skip(self, word):
+        return (word == word.upper() # all caps
+                or
+                # pluralized acronym ("URLs")
+                (word[-1].lower() == 's'
+                 and
+                 word[:-1] == word[:-1].upper()
+                 )
+                )
+
+class list_tokenize(tokenize):
+    def __init__(self, words):
+        tokenize.__init__(self, '')
+        self._words = words
+    def next(self):
+        if not self._words:
+            raise StopIteration()
+        word = self._words.pop(0)
+        return (word, 0)
+
+class ContractionFilter(Filter):
+    """Strip common contractions from words.
+    """
+    splits = {
+        "won't":['will', 'not'],
+        "isn't":['is', 'not'],
+        "can't":['can', 'not'],
+        "i'm":['I', 'am'],
+        }
+    def _split(self, word):
+        # Fixed responses
+        if word.lower() in self.splits:
+            return list_tokenize(self.splits[word.lower()])
+
+        # Possessive
+        if word.lower().endswith("'s"):
+            return unit_tokenize(word[:-2])
+
+        # * not
+        if word.lower().endswith("n't"):
+            return unit_tokenize(word[:-3])
+
+        return unit_tokenize(word)
+
+class IgnoreWordsFilter(Filter):
+    """Given a set of words, ignore them all.
+    """
+    def __init__(self, tokenizer, word_set):
+        self.word_set = set(word_set)
+        Filter.__init__(self, tokenizer)
+    def _skip(self, word):
+        return word in self.word_set
+
+class IgnoreWordsFilterFactory(object):
+    def __init__(self, words):
+        self.words = words
+    def __call__(self, tokenizer):
+        return IgnoreWordsFilter(tokenizer, self.words)
+
+class PyPIFilterFactory(IgnoreWordsFilterFactory):
+    """Build an IgnoreWordsFilter for all of the names of packages on PyPI.
+    """
+    def __init__(self):
+        client = xmlrpclib.ServerProxy('http://pypi.python.org/pypi')
+        IgnoreWordsFilterFactory.__init__(self, client.list_packages())
+
+class PythonBuiltinsFilter(Filter):
+    """Ignore names of built-in Python symbols.
+    """
+    def _skip(self, word):
+        return word in __builtins__
+
+class ImportableModuleFilter(Filter):
+    """Ignore names of modules that we could import.
+    """
+    def __init__(self, tokenizer):
+        Filter.__init__(self, tokenizer)
+        self.found_modules = set()
+        self.sought_modules = set()
+    def _skip(self, word):
+        if word not in self.sought_modules:
+            self.sought_modules.add(word)
+            try:
+                imp.find_module(word)
+            except UnicodeEncodeError:
+                return False
+            except ImportError:
+                return False
+            else:
+                self.found_modules.add(word)
+                return True
+        return word in self.found_modules
+
 
 class SpellingChecker(object):
     """Checks the spelling of blocks of text.
@@ -32,15 +170,26 @@ class SpellingChecker(object):
     Uses options defined in the sphinx configuration file to control
     the checking and filtering behavior.
     """
-    
-    def __init__(self, lang, suggest, word_list_filename):
-        # TODO - filters
+
+    def __init__(self, lang, suggest, word_list_filename, filters=[]):
         self.dictionary = enchant.DictWithPWL(lang, word_list_filename)
-        self.tokenizer = get_tokenizer(lang)
+        self.tokenizer = get_tokenizer(lang, filters)
+        self.original_tokenizer = self.tokenizer
         self.suggest = suggest
-        
-    _class_name_pattern = re.compile(r'[A-Z]+[a-z0-9]+[A-Z]+[a-z0-9]+')
-    _spelling_word_pattern = re.compile(r'\w+', re.UNICODE|re.MULTILINE|re.DOTALL)
+
+    def push_filters(self, new_filters):
+        """Add a filter to the tokenizer chain.
+        """
+        t = self.tokenizer
+        for f in new_filters:
+            t = f(t)
+        self.tokenizer = t
+
+    def pop_filters(self):
+        """Remove the filters pushed during the last call to push_filters().
+        """
+        self.tokenizer = self.original_tokenizer
+
     def check(self, text):
         """Generator function that yields bad words and suggested alternate spellings.
         """
@@ -48,174 +197,16 @@ class SpellingChecker(object):
             correct = self.dictionary.check(word)
             if correct:
                 continue
-
-            # TODO - filter to ignore acronyms
-#             if word.upper() == word:
-#                 # assume acronyms are right
-#                 continue
-#             if word.rstrip('s').upper() == word.rstrip('s'):
-#                 # plural acronyms, too
-#                 continue
-
-            # TODO - filter to ignore python builtins
-#             if word in __builtins__:
-#                 # python built-in
-#                 continue
-
-            # TODO - filter to ignore "code-like" patterns, maybe via an option defining regex?
-#             if '_' in word:
-#                 # probably something from code
-#                 continue
-
-            # TODO - option to ignore importable modules
-#             try:
-#                 imp.find_module(word)
-#             except ImportError:
-#                 pass
-#             else:
-#                 # importable module name
-#                 try:
-#                     SPELLER.addtoSession(word)
-#                 except aspell.AspellSpellerError:
-#                     pass
-#                 continue
-
-            # TODO - option to include suggestions
             yield word, self.dictionary.suggest(word) if self.suggest else []
-
         return
 
-class Section(object):
-    
-    def __init__(self, spelling_checker):
-        self.title = ''
-        self.spelling_checker = spelling_checker
-        self.errors = []
-        self.subsections = []
-        
-    def get_text_lines(self):
-        lines = self.errors[:]
 
-        for subsection in self.subsections:
-            lines.extend(' ' + line
-                         for line in subsection.get_text_lines()
-                         )
-
-        if lines:
-            lines.insert(0, self.title)
-            lines.insert(0, '')
-        return lines
-    
-    def add_subsection(self, subsection):
-        self.subsections.append(subsection)
-        return
-    
-    def set_title(self, title):
-        self.title = title
-        self.check_spelling(title)
-        
-    def astext(self):
-        lines = itertools.chain(*tuple(section.get_text_lines()
-                                       for section in self.subsections
-                                       )
-                                 )
-        # May want this formatting when suggestions are added
-        real_lines = []
-        for line in lines:
-            initial_indent = line[:len(line)-len(line.lstrip())]
-            if line.lstrip().startswith('-'): # error line
-                subsequent_indent = initial_indent + '  '
-            else:
-                subsequent_indent = initial_indent
-            line = textwrap.fill(line.lstrip(), width=70,
-                                 initial_indent=initial_indent,
-                                 subsequent_indent=subsequent_indent,
-                                 )
-            real_lines.append(line)
-        return '\n'.join(real_lines)
-
-    def add_text(self, text):
-        self.check_spelling(text)
-    
-    def check_spelling(self, text):
-        for word, suggestions in self.spelling_checker.check(text):
-            msg = '- "%s"' % word
-            if suggestions:
-                msg = '%s: %s' % (msg, ', '.join(suggestions))
-            self.errors.append(msg)
-        return
-        
-TEXT_NODES = set([ 'block_quote', 'paragraph',
-                   'list_item', 'term', 'definition_list_item', ])
-    
-
-class SpellingVisitor(GenericNodeVisitor):
-    """A Visitor class; see the docutils for more details.
-    """
-    def __init__(self, checker, *args, **kw):
-        GenericNodeVisitor.__init__(self, *args, **kw)
-        self.spelling_checker = checker
-        self.sections = [ Section(self.spelling_checker) ]
-
-    def visit_title(self, node):
-        # Set the title for the current section.
-        if not self.sections[-1].title:
-            self.sections[-1].set_title(node.astext())
-
-    def visit_section(self, node):
-        # Start a new section
-        new_section = Section(self.spelling_checker)
-        self.sections[-1].add_subsection(new_section)
-        self.sections.append(new_section)
-        
-    def depart_section(self, node):
-        # Pop the section stack
-        self.sections.pop()
-
-    def visit_Text(self, node):
-        if node.parent.tagname in TEXT_NODES:
-            self.sections[-1].add_text(node.astext())
-
-    def default_visit(self, node): pass
-    def default_departure(self, node): pass
-
-    def unknown_visit(self, node):
-        self.document.reporter.warning('Ignoring node: %s' % node.tagname)
-    def unknown_departure(self, node): pass
-
-    # Ignore conditional "only" nodes
-    def visit_only(self, node): pass
-    def depart_only(self, node): pass
-
-    # Ignore inline literal text
-    def visit_literal(self, node): pass
-    def depart_literal(self, node): pass
-    def visit_emphasis(self, node): pass
-    def depart_emphasis(self, node): pass
-
-    # Ignore comment blocks
-    def visit_comment(self, node): pass
-    def depart_comment(self, node): pass
-
-    # Ignore literal blocks
-    def visit_literal_block(self, node): pass
-
-    def astext(self):
-        body = '\n'.join(section.astext()
-                         for section in self.sections
-                         )
-        return body
-
-
-class SpellingWriter(Writer):
-    """Boilerplate attaching our Visitor to a docutils document."""
-    def __init__(self, checker, *args, **kwds):
-        self.spelling_checker = checker
-        Writer.__init__(self, *args, **kwds)
-    def translate(self):
-        visitor = SpellingVisitor(self.spelling_checker, self.document)
-        self.document.walkabout(visitor)
-        self.output = '\n' + visitor.astext() + '\n'
+TEXT_NODES = set([ 'block_quote',
+                   'paragraph',
+                   'list_item',
+                   'term',
+                   'definition_list_item',
+                   ])
 
 
 class SpellingBuilder(Builder):
@@ -228,44 +219,113 @@ class SpellingBuilder(Builder):
         self.docnames = []
         self.document_data = []
 
+        filters = [ ContractionFilter,
+                    EmailFilter,
+                    ]
+        if self.config.spelling_ignore_wiki_words:
+            filters.append(WikiWordFilter)
+        if self.config.spelling_ignore_acronyms:
+            filters.append(AcronymFilter)
+        if self.config.spelling_ignore_pypi_package_names:
+            self.info('Adding package names from PyPI to local spelling dictionary...')
+            filters.append(PyPIFilterFactory())
+        if self.config.spelling_ignore_python_builtins:
+            filters.append(PythonBuiltinsFilter)
+        if self.config.spelling_ignore_importable_modules:
+            filters.append(ImportableModuleFilter)
+        filters.extend(self.config.spelling_filters)
+
+        project_words = os.path.join(self.srcdir, self.config.spelling_word_list_filename)
+        self.checker = SpellingChecker(lang=self.config.spelling_lang,
+                                       suggest=self.config.spelling_show_suggestions,
+                                       word_list_filename=project_words,
+                                       filters=filters,
+                                       )
+        self.output_filename = os.path.join(self.outdir, 'output.txt')
+        self.output = codecs.open(self.output_filename, 'wt', encoding='UTF-8')
+
     def get_outdated_docs(self):
         return 'all documents'
 
-    def write(self, *ignored):
-        self.output = []
-
-        checker = SpellingChecker(lang=self.config.spelling_lang,
-                                  suggest=self.config.spelling_show_suggestions,
-                                  word_list_filename=self.config.spelling_word_list_filename,
-                                  )
-        
-        master = self.config.master_doc
-        docwriter = SpellingWriter(checker=checker)
-        docsettings = OptionParser(
-            defaults=self.env.settings,
-            components=(docwriter,)).get_default_values()
-            
-        tree = self.env.get_doctree(master)
-        tree = inline_all_toctrees(self, set(), master, tree, darkgreen)
-        tree['docname'] = master
-
-        destination = StringOutput(encoding='utf-8')
-        text = docwriter.write(tree, destination).strip()
-        if text:
-            self.output.append(text)
-
+    def prepare_writing(self, docnames):
         return
-    
+
+    def get_target_uri(self, docname, typ=None):
+        return ''
+
+    def format_suggestions(self, suggestions):
+        if not self.config.spelling_show_suggestions or not suggestions:
+            return u''
+        return u'[' + u', '.join(u'"%s"' % s for s in suggestions) + u']'
+
+    def write_doc(self, docname, doctree):
+        self.checker.push_filters(self.env.spelling_document_filters[docname])
+        
+        for node in doctree.traverse(docutils.nodes.Text):
+            if node.tagname == '#text' and  node.parent.tagname in TEXT_NODES:
+
+                # Figure out the line number for this node by climbing the
+                # tree until we find a node that has a line number.
+                lineno = None
+                parent = node
+                seen = set()
+                while lineno is None:
+                    #self.info('looking for line number on %r' % node)
+                    seen.add(parent)
+                    parent = node.parent
+                    if parent is None or parent in seen:
+                        break
+                    lineno = parent.line
+                filename = self.env.doc2path(docname, base=None)
+
+                # Check the text of the node.
+                for word, suggestions in self.checker.check(node.astext()):
+                    msg_parts = []
+                    if lineno:
+                        msg_parts.append(darkgreen('(line %3d)' % lineno))
+                    msg_parts.append(red(word))
+                    msg_parts.append(self.format_suggestions(suggestions))
+                    msg = ' '.join(msg_parts)
+                    self.info(msg)
+                    self.output.write(u"%s:%s: (%s) %s\n" % (
+                            self.env.doc2path(docname, None),
+                            lineno, word,
+                            self.format_suggestions(suggestions),
+                            ))
+
+                    # We found at least one bad spelling, so set the status
+                    # code for the app to a value that indicates an error.
+                    self.app.statuscode = 1
+
+        self.checker.pop_filters()
+        return
+
     def finish(self):
-        self.info()
-        for text in self.output:
-            self.info(text)
+        self.output.close()
+        self.info('Spelling checker messages written to %s' % self.output_filename)
         return
 
 def setup(app):
     app.info('Initializing Spelling Checker')
     app.add_builder(SpellingBuilder)
+    # Report guesses about correct spelling
     app.add_config_value('spelling_show_suggestions', False, 'env')
+    # Set the language for the text
     app.add_config_value('spelling_lang', 'en_US', 'env')
+    # Set a user-provided list of words known to be spelled properly
     app.add_config_value('spelling_word_list_filename', 'spelling_wordlist.txt', 'env')
+    # Assume anything that looks like a PyPI package name is spelled properly
+    app.add_config_value('spelling_ignore_pypi_package_names', False, 'env')
+    # Assume words that look like wiki page names are spelled properly
+    app.add_config_value('spelling_ignore_wiki_words', True, 'env')
+    # Assume words that are all caps, or all caps with trailing s, are spelled properly
+    app.add_config_value('spelling_ignore_acronyms', True, 'env')
+    # Assume words that are part of __builtins__ are spelled properly
+    app.add_config_value('spelling_ignore_python_builtins', True, 'env')
+    # Assume words that look like the names of importable modules are spelled properly
+    app.add_config_value('spelling_ignore_importable_modules', True, 'env')
+    # Add any user-defined filter classes
+    app.add_config_value('spelling_filters', [], 'env')
+    # Register the 'spelling' directive for setting parameters within a document
+    rst.directives.register_directive('spelling', SpellingDirective)
     return