+++ issuetracker/CHANGES.rst
@@ -1,3 +1,11 @@
+0.7.2 (Mar 10, 2011)
+====================
+
+- #13: Fixed source distribution to include tests again
+- Fixed extraction of issue state for open issues from bitbucket
+- Ignore references in inline literals and literal blocks
+
+
 0.7.1 (Jan 19, 2011)
 ====================